Better Layer constructor, template providers
The `MM.Layer` constructor accepts an optional 3rd argument: the parent
element, which is only added to the DOM if it doesn't have a parentNode
yet.

Merged in some improvements to `MM.TemplatedMapProvider`, which now
understands Microsoft-style templates and does quadkey replacements. You
can also provide subdomain replacements in the form: "{S:1,2,3}".