Reversed tile sizing behavior
My hunch about forcing CSS width and height on tile elements inside
MM.moveElement() was wrong, and actually broke a lot of stuff. So I
reversed it and fixed the sprite-tiles example layer class to give each
tile an intrinsic size with a block-level "strut", which appears to work
well.

Also, the spotlight example has a nicer transition to it with a slight
delay.